swhich process on install pyenv

+++ pyenv-install.sh
@@ -1,4 +1,8 @@
-sudo apt-get install -y libssl-dev libbz2-dev libreadline-dev libsqlite3-dev libffi-dev
+if [ -f "/etc/redhat-release" ]; then
+  sudo yum install -y openssl-devel bzip2-devel readline-devel sqlite-devel libffi-devel
+elif [ -f "/etc/lsb-release" ]; then
+  sudo apt-get install -y libssl-dev libbz2-dev libreadline-dev libsqlite3-dev libffi-dev
+fi
 
 # latest python 2.7.x
 PYTHON2=`pyenv install --list | grep -E "^ +2.7." | grep -v "dev" | awk 'END{gsub(" ", ""); print}'`
